App and main refactored
