+++ main.rb
@@ -2,7 +2,9 @@ require_relative 'app'
 
 def main
   app = App.new
-  app.run
+  loop do
+    app.run
+  end
 end
 
 main
